requirements.txt
anyio==4.3.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0.2
fonttools==4.49.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-telegram-bot==20.8
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
